Created a ui button scrit for building ui

+++ Assets/Scripts/Buildings/BuildingCost.cs
@@ -4,9 +4,9 @@ namespace Buildings {
     [System.Serializable]
     public readonly struct BuildingCost {
         public readonly ResourceType resourceType;
-        public readonly int amount;
+        public readonly float amount;
 
-        public BuildingCost(ResourceType resourceType, int amount) {
+        public BuildingCost(ResourceType resourceType, float amount) {
             this.resourceType = resourceType;
             this.amount = amount;
         }
